api/requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
Werkzeug==2.2.3
